Wave tracking for enemy spawners
Enemy spawners now support waves with specified numbers of each enemy. Waves are stored in the WaveTally script attached to the core. Each wave is stored as a vector3, where the X coordinate refers to the number of fastbots to spawn, the Y coordinate refers to the number of slowbots to spawn, and the Z coordinate refers to the number of rangebots to spawn. NOTE: the public variables "wave" and "waveDone" on the WaveTally script are public only so that other scripts can access them, and should not be edited manually.