Added wave completion
Once all enemies in a wave have been killed, the core will now set a 20 second timer (this amount is a public variable so it can be changed) after which the next wave will begin. Each wave completion triggers the core controller's endWave() function, which currently just prints text to the debug log but in the future will be able to do things like give the player a point bonus or send a message to the UI. Endless mode is still supported, but it complicates the code so it may be removed in the future.