Created token manager script
Token manager script should be attached to the main camera in order to work properly. It allows upgrading turrets and repairing the core with right-click, and once weapon upgrading exists it will be triggered by pressing the E key. NOTE: collection of in-world tokens is not handled by this script, and should take place in a script connected to the player's hitbox rather than one connected to the main camera.